- disabled console and onscreen keyboard temporarily to make a release - temporary fix for xaudio2 sound delay problem - some x64 fixes (added the avp linux fixes)
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@167

+++ 3dc/avp/bh_plachier.c
@@ -429,7 +429,8 @@ void SaveStrategy_PlacedHierarchy(STRATEGYBLOCK* sbPtr)
 	COPY_NAME(header->SBname,sbPtr->SBname);
 
 	{
-		int sequence_index = ph_bhv->current_seq-ph_bhv->sequences;
+//		int sequence_index = ph_bhv->current_seq-ph_bhv->sequences;
+		intptr_t sequence_index = ph_bhv->current_seq-ph_bhv->sequences; // x64
 		*(int*)buffer = sequence_index;
 		buffer+=sizeof(int);
 	}
